Move utils lib to poetry package package
