--- batch-processing/libs/batch_utils/README.md
@@ -1,1 +1,0 @@
-# Batch utils
