--- batch-processing/libs/batch_utils/batch_utils/__init__.py
@@ -1,10 +1,0 @@
-from .batch_utils import (
-    date_to_mjd,
-    str_to_date,
-    dates_between_generator,
-    configure_logger,
-    _init_path,
-    _rm_directory_or_file,
-    path_exists,
-    drop_polars_columms,
-)
