--- batch-processing/libs/batch_utils/batch_utils/prv_candidates/__init__.py
@@ -1,4 +1,0 @@
-from .step.prv_candidates import extract_detections_and_non_detections_dataframe
-from .step.prv_candidates import extract_detections_and_non_detections_dataframe_multiprocessing
-if __name__ == "__main__":
-    print(extract_detections_and_non_detections_dataframe)
