--- batch-processing/libs/batch_utils/batch_utils/sorting_hat/__init__.py
@@ -1,5 +1,0 @@
-from .step.sorting_hat import df_sorting_hat
-from .step.sorting_hat import sorting_hat_aid_replacer
-
-if __name__ == "__main__":
-    print(df_sorting_hat)
