--- batch-processing/libs/batch_utils/batch_utils/types.py
@@ -1,5 +1,0 @@
-from enum import auto, Enum
-
-class ParquetFolder(Enum):
-    RawParquet = auto()
-    SortingHatParquet = auto()
