--- batch-processing/libs/batch_utils/pyproject.toml
@@ -1,13 +1,0 @@
-[tool.poetry]
-name = "batch-utils"
-version = "0.1.0"
-description = "Common methods for batch processing"
-authors = ["Edison Pizarro <edison.pizarro.godoy@gmail.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
